fix: validate bwrb vault on startup
